# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma de empleo ProviEmplea. Permite gestionar talentos, empresas y procesos de selección con curriculum ciego.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Listar talentos en formato CV ciego",
                "description": "Retorna perfiles sin datos personales identificables. Resultados recomendados para caché del lado cliente por 5 minutos. Límite: 60 requests/minuto.",
                "operationId": "getPersonas",
                "parameters": [
                    {
                        "name": "validado",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "nivel_educacional",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "basica",
                                "media",
                                "tecnica",
                                "universitaria",
                                "postgrado"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Persona"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiadas solicitudes. Límite: 60 req/min. Espere el tiempo indicado en el header Retry-After."
                    }
                }
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Registrar nuevo talento",
                "operationId": "createPersona",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "talento@ejemplo.cl"
                                    },
                                    "nivel_educacional": {
                                        "type": "string",
                                        "enum": [
                                            "basica",
                                            "media",
                                            "tecnica",
                                            "universitaria",
                                            "postgrado"
                                        ]
                                    },
                                    "anios_experiencia": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "tipo_jornada": {
                                        "type": "string",
                                        "enum": [
                                            "completa",
                                            "part-time",
                                            "por-horas"
                                        ]
                                    },
                                    "modalidad": {
                                        "type": "string",
                                        "enum": [
                                            "presencial",
                                            "remoto",
                                            "hibrido"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/personas/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener persona por ID",
                "operationId": "getPersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualizar persona",
                "operationId": "updatePersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona actualizada"
                    },
                    "404": {
                        "description": "No encontrada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Desactivar persona",
                "operationId": "deletePersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona desactivada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        },
        "/personas/{id}/validar": {
            "patch": {
                "tags": [
                    "Personas"
                ],
                "summary": "Validar perfil de persona",
                "operationId": "validarPersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona validada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Persona": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "codigo_talento": {
                        "type": "string",
                        "example": "PROV-2026-A1B2"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "nivel_educacional": {
                        "type": "string",
                        "nullable": true
                    },
                    "anios_experiencia": {
                        "type": "integer",
                        "example": 3
                    },
                    "tipo_jornada": {
                        "type": "string",
                        "nullable": true
                    },
                    "modalidad": {
                        "type": "string",
                        "nullable": true
                    },
                    "validado": {
                        "type": "boolean",
                        "example": false
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "porcentaje_completitud": {
                        "type": "integer",
                        "example": 73
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "rateLimitHeaders": {
                "type": "apiKey",
                "name": "X-RateLimit-Limit",
                "in": "header"
            }
        }
    },
...
}